+++ RELEASE_NOTES.md
@@ -19,6 +19,7 @@
 ### Cookiecutter template
 
 * Group GitHub Actions dependabot updates.
+* API projects don't include the `google-common-protos` dependency by default.
 
 ## Bug Fixes